refactor scan type
