+++ zap/src/zap_scan_type.py
@@ -1,0 +1,24 @@
+"""
+Provides the ScanType Enum.
+"""
+
+from enum import Enum
+
+
+class ScanType(str, Enum):
+    """
+    Enumerates Zap compliance scan types
+    """
+
+    API = "API"
+    AUTH = "Authenticated"
+    BASELINE = "Baseline"
+    UI = "UI"
+    LEOAPP = "LeoApp"
+
+    def __str__(self):
+        return str(self.name).lower()
+
+    def label(self):
+        """"Get user-friendly name of the scan type"""
+        return str(self.value)
